Refactored stone base classes to reduce code duplication that may result in disparity between behavior
