+++ TFC_Shared/src/TFC/Blocks/Terrain/BlockSed.java
@@ -23,48 +23,14 @@ import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 
 
-public class BlockSed extends BlockCollapsable
+public class BlockSed extends BlockStone
 {
     public BlockSed(int i, Material material, int id) {
         super(i, material, id);
         this.dropBlock = TFCBlocks.StoneSedCobble.blockID;
-    }
 
-    @SideOnly(Side.CLIENT)
-    @Override
-    /**
-     * returns a list of blocks with the same ID, but different meta (eg: wood returns 4 blocks)
-     */
-    public void getSubBlocks(int par1, CreativeTabs par2CreativeTabs, List par3List)
-    {
-    	for(int i = 0; i < 10; i++)
-    		par3List.add(new ItemStack(par1, 1, i));
-    }
-
-    /*
-     * Mapping from metadata value to damage value
-     */
-    @Override
-    public int damageDropped(int i) {
-        return i;
-    }
-
-    @Override
-	public Icon getIcon(int i, int j) 
-	{
-		return icons[j];
-	}
-
-    public static Icon[] icons = new Icon[10];
-    public String[] names = {"Siltstone", "Mudstone", "Shale", "Claystone", "Rock Salt", "Limestone", "Conglomerate", "Dolomite", "Chert", "Chalk"};
-	
-	@Override
-	public void registerIcons(IconRegister iconRegisterer)
-    {
-		for(int i = 0; i < 10; i++)
-		{
-			icons[i] = iconRegisterer.registerIcon("rocks/"+names[i]+" Raw");
-		}
+        names = new String[] {"Siltstone", "Mudstone", "Shale", "Claystone", "Rock Salt", "Limestone", "Conglomerate", "Dolomite", "Chert", "Chalk"};
+        icons = new Icon[names.length];
     }
 
     @Override
@@ -75,32 +41,6 @@ public class BlockSed extends BlockCollapsable
             dropBlockAsItem_do(world, i, j, k, new ItemStack(TFCItems.LooseRock, R.nextInt(4), l+3));
 
         super.harvestBlock(world, entityplayer, i, j, k, l);
-    }
-
-    @Override
-    public int idDropped(int i, Random random, int j)
-    {
-        return TFCItems.LooseRock.itemID;
-    }
-
-    @Override
-	public void onBlockDestroyedByExplosion(World world, int i, int j, int k, Explosion ex) 
-    {
-        if(!world.isRemote)
-        {
-            //super.onBlockDestroyedByExplosion(world, i, j, k, ex);
-            Random random = new Random();
-
-            ItemStack is = null;
-
-            is = TFC_Core.RandomGem(random, 0);
-
-            if(is != null)
-            {
-                EntityItem item = new EntityItem(world, i, j, k, is);
-                world.spawnEntityInWorld(item);
-            }
-        }
     }
 
     @Override
@@ -121,12 +61,6 @@ public class BlockSed extends BlockCollapsable
             }
 
         }
-    }
-
-    @Override
-	public void onNeighborBlockChange(World world, int i, int j, int k, int l)
-    {
-        DropCarvedStone(world, i, j, k);
     }
     
     /**
